MINOR FIX FOR DEPLOY

+++ database/migrations/2019_03_15_152822_create_revisions_table.php
@@ -17,7 +17,7 @@ class CreateRevisionsTable extends Migration
             $table->bigIncrements('id');
             $table->string('filename');
             $table->string('notes');
-            $table->smallInteger('attcomp');
+            $table->float('attcomp');
             $table->unsignedBigInteger('excel_id');
             $table->foreign('excel_id')->references('id')->on('excels');
             $table->unsignedBigInteger('user_id');
